requirements-api.txt
# =============================================================================
# API Dependencies
# =============================================================================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# ML Inference (CPU-only PyTorch)
--extra-index-url https://download.pytorch.org/whl/cpu
torch
torchvision
ultralytics>=8.1.0

# MLflow (for model loading)
mlflow>=2.10.0
boto3>=1.34.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0

# Monitoring (optional)
prometheus-client>=0.19.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
